Uhm, works, but zlib is flaky, and I'm shelling out to tar
I have to read the whole file in w/ zlib,
otherwise it locks the process trying to read the file or something.
If I use .read, it works correctly.
Kinda shitty, but w/e, I can deal with it if I maintain this.

There's a tar lib that was edited in December
https://github.com/halostatue/minitar (I think I met that dude at Rubyconf!)

But the gem hasn't been updated since 2008
https://rubygems.org/gems/archive-tar-minitar

So, going to shell out for now, just b/c it's easier than figuring that out.